rename ComponentResource to Components
